Address review feedback: fix overflow risk, remove unused variables
- Use Uint64 arithmetic in SD_GetSubTickFraction() to prevent overflow
  when extra_ms * 70 exceeds 32-bit range
- Remove unused rf_smoothScrollDX/DY variables and their assignments

Agent-Logs-Url: https://github.com/awe444/BioMenaceDecomp/sessions/5e0a9050-56c6-46b2-b443-df58226c28f1

+++ SRC/ID_SD.c
@@ -188,7 +188,7 @@ SD_GetSubTickFraction(void)
 {
   Uint32  now = SDL_GetTicks();
   Uint32  extra_ms = now - sdl_lastTicks;
-  Uint32  total = sdl_tickRemainder + extra_ms * 70;
+  Uint64  total = (Uint64)sdl_tickRemainder + (Uint64)extra_ms * 70;
   double  frac = (double)total / 1000.0;
 
   if (frac < 0.0) frac = 0.0;
